test(c): migrate test_validate_core.py -> integration/test_validate.py (#40)
Four tests for validate.run_validate. Classes flattened per Rule M2.
FakeFailingSource patch preserved in connection-failure test.
