--- tests/test_validate_core.py
@@ -1,93 +1,0 @@
-"""Direct unit tests for feather_etl.validate.run_validate()."""
-
-from __future__ import annotations
-
-import shutil
-from pathlib import Path
-
-import yaml
-
-from tests.conftest import FIXTURES_DIR
-from tests.helpers import make_curation_entry, write_curation
-
-
-def _make_sqlite_project(tmp_path: Path, *, valid_path: bool = True) -> Path:
-    """Build a minimal feather project with a SQLite source. Returns config path."""
-    if valid_path:
-        shutil.copy2(FIXTURES_DIR / "sample_erp.sqlite", tmp_path / "source.sqlite")
-        source_path = "./source.sqlite"
-    else:
-        source_path = "./missing.sqlite"
-
-    config = {
-        "sources": [{"type": "sqlite", "name": "erp", "path": source_path}],
-        "destination": {"path": "./feather_data.duckdb"},
-    }
-    config_path = tmp_path / "feather.yaml"
-    config_path.write_text(yaml.dump(config))
-    write_curation(
-        tmp_path,
-        [make_curation_entry("erp", "orders", "orders")],
-    )
-    return config_path
-
-
-class TestRunValidate:
-    def test_reports_each_source_status_when_all_ok(self, tmp_path: Path):
-        from feather_etl.config import load_config
-        from feather_etl.validate import run_validate
-
-        cfg = load_config(_make_sqlite_project(tmp_path))
-        report = run_validate(cfg)
-
-        assert report.all_ok is True
-        assert report.tables_count == 1
-        assert len(report.sources) == 1
-        assert report.sources[0].type == "sqlite"
-        assert report.sources[0].ok is True
-        assert report.sources[0].error is None
-
-    def test_returns_all_ok_false_when_any_source_check_fails(self, tmp_path: Path):
-        from feather_etl.config import load_config
-        from feather_etl.validate import run_validate
-
-        cfg = load_config(_make_sqlite_project(tmp_path, valid_path=False))
-        report = run_validate(cfg)
-
-        assert report.all_ok is False
-        assert report.sources[0].ok is False
-
-    def test_propagates_last_error_for_failed_sources(self, tmp_path: Path):
-        from feather_etl.config import load_config
-        from feather_etl.validate import run_validate
-
-        cfg = load_config(_make_sqlite_project(tmp_path))
-
-        # Replace the source with a fake that fails with a diagnostic error —
-        # mirrors the real-world case (DB sources set `_last_error` on failure)
-        # that the `validate` command surfaces as "Details: ...".
-        class FakeFailingSource:
-            type = "duckdb"
-            path = "/nope.duckdb"
-
-            def __init__(self) -> None:
-                self._last_error = "TLS Provider: certificate verify failed"
-
-            def check(self) -> bool:
-                return False
-
-        cfg.sources[0] = FakeFailingSource()
-        report = run_validate(cfg)
-
-        assert report.sources[0].ok is False
-        assert report.sources[0].error == "TLS Provider: certificate verify failed"
-
-    def test_label_uses_path_for_file_sources(self, tmp_path: Path):
-        from feather_etl.config import load_config
-        from feather_etl.validate import run_validate
-
-        cfg = load_config(_make_sqlite_project(tmp_path))
-        report = run_validate(cfg)
-
-        # File-based sources expose `path` — the label should be the path.
-        assert "source.sqlite" in str(report.sources[0].label)
